ref: rename SentryViewController to top-level func
Follow-up on #8491. Renames the SentryViewController enum namespace to a
top-level viewControllerHierarchy(from:) function, because
SentryViewController is a quite confusing name.

#skip-changelog
