+++ Sources/Swift/Core/Integrations/Performance/SentryUIViewControllerSwizzling.swift
@@ -234,7 +234,7 @@ class SentryUIViewControllerSwizzling {
     }
 
     func swizzleRootViewControllerAndDescendant(_ rootViewController: UIViewController) {
-        let allViewControllers = SentryViewController.descendants(of: rootViewController)
+        let allViewControllers = viewControllerHierarchy(from: rootViewController)
 
         SentrySDKLog.debug("Found \(allViewControllers.count) descendants for RootViewController \(rootViewController)")
